# openapi.yml
openapi: 3.1.0
info:
  title: letsgaming.de backend API
  version: 1.1.0
  license:
    name: MIT
  summary: The Fastify API behind letsgaming.de — a public read/relay surface and an authenticated CMS.
  description: |
    The backend for [letsgaming.de](https://letsgaming.de): a small Fastify service that
    accumulates data from GitHub/Steam/Wakapi/Lanyard into a local SQLite store and exposes
    it to the site.

    ## The one architectural rule: the server is the boundary
    The browser never talks to a third party directly. Discord, Steam and Spotify images,
    Lanyard presence, and every other upstream are fetched **server-side** and returned
    already filtered, so a visitor only ever contacts this origin. This is why endpoints like
    `/api/presence` and `/api/presence/media` exist at all — they are relays, not passthroughs.

    ## Two surfaces, two threat models
    - **Public** (`/api/*`, `/assets/*`, `/health`) — unauthenticated reads and a
      few write **relays** (contact, guestbook, engagement). Writes here persist only counters
      or nothing at all, are rate-limited per-IP in memory, and never set a cookie.
    - **CMS** (`/api/cms/*`) — authenticated content editing. A different threat model, kept on
      its own path; session/auth concerns never leak into the public surface.

    The site's own pages do **not** call `/api/site` at runtime — SSR reads the SQLite store
...
    app's fallback when it can't open the store file.

...

  contact:
...
servers:
...

...
security:
  - cmsSession: []
  - cmsBearer: []
...
paths:
...
  /api/cms/status:
    put:
      tags: [CMS · Content]
      summary: Update the "currently building" status (verb + now)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HeroStatusBody" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    get:
      tags: [CMS · Status]
      summary: Dashboard status
      description: >
        Per-job sync status (sources, presence sampler, metadata sweeps, analytics ingest), guestbook
        counts by status, and the ten most recent content edits.
      responses:
        "200":
          description: Status snapshot.
          content:
            application/json:
              schema:
                type: object
                required: [sources, guestbook, recentEdits]
                properties:
                  sources:
                    type: array
                    description: >
                      Always lists github, wakapi, presence, game-metadata, game-images and analytics,
                      including never-synced and unconfigured ones.
                    items:
                      type: object
                      required: [id, label, kind, configured, mock, schedule, canSync, state, lastSuccessAt, lastErrorAt, lastError]
                      properties:
                        id: { type: string }
                        label: { type: string }
                        kind: { type: string, enum: [source, job, analytics] }
                        configured: { type: boolean }
                        mock: { type: boolean, description: Running on the deterministic mock source. }
                        schedule: { type: [string, "null"], description: "Cron expression, sources only." }
                        canSync: { type: boolean, description: "True when POST /api/cms/sync/{source} accepts this id." }
                        state: { type: string, enum: [ok, error, never], description: "`error` when the latest attempt failed." }
                        lastSuccessAt: { type: [string, "null"], format: date-time }
                        lastErrorAt: { type: [string, "null"], format: date-time }
                        lastError: { type: [string, "null"] }
                  guestbook:
                    type: object
                    required: [pending, approved, rejected]
                    properties:
                      pending: { type: integer }
                      approved: { type: integer }
                      rejected: { type: integer }
                  recentEdits:
                    type: array
                    items:
                      type: object
                      required: [id, savedAt, reason, label, kind, restorable]
                      properties:
                        id: { type: integer, description: "Revision id within the id space named by `kind`. Only `content` ids are valid for /api/cms/revisions." }
                        kind: { type: string, enum: [content, ia], description: "`ia` rows are layout and section-heading edits, archived separately." }
                        restorable: { type: boolean, description: "True only for `content` rows." }
                        savedAt: { type: string, format: date-time }
                        reason: { type: string, description: "Raw table/column name, `restore:<id>`, or `ia:<nav|modules|module-added|module-removed>` for layout edits." }
                        label: { type: string, description: Human-readable label for `reason`. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/cms/sync/{source}:
    post:
      tags: [CMS · Status]
      summary: Sync one source now
      description: >
        Runs a registered source immediately and records the outcome. A trigger that arrives while the
        same source is already syncing joins that run and returns its result.
      parameters:
        - { name: source, in: path, required: true, schema: { type: string, enum: [github, wakapi] } }
      responses:
        "200":
          description: The sync result. A failed sync is still 200 with `ok:false` and `error`.
          content:
            application/json:
              schema:
                type: object
                required: [sourceId, ok, mock, syncedAt]
                properties:
                  sourceId: { type: string }
                  ok: { type: boolean }
                  mock: { type: boolean }
                  syncedAt: { type: string, format: date-time }
                  error: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { description: The sync worker is not running. }
  /api/cms/sections:
    get:
      tags: [CMS · Status]
      summary: Section stats for the editor overlay
      description: >
        Per-nav-node stats from the in-page tracker only (`source: script`). Tracking granularity is
        the nav node, so modules inside one area share that area's numbers (`modules` lists them).
        Median dwell is approximated from the dwell buckets as the midpoint of the bucket holding the
        50th percentile (<5s 2.5s, 5-15s 10s, 15-30s 22.5s, 30-60s 45s, 1-3m 120s, 3-10m 390s,
        10m+ 900s). Reach is the share of completed visits whose scroll reached 25% of the section,
        capped at 1; null when there are no completed visits in the window.
      parameters:
        - { name: area, in: query, required: false, schema: { type: string }, description: Nav node id; returns it and its descendants. Omit for every node. }
        - { name: hours, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 8760, default: 168 } }
      responses:
        "200":
          description: Section stats.
          content:
            application/json:
              schema:
                type: object
                required: [source, hours, from, to, visits, sections]
                properties:
                  source: { type: string, enum: [script] }
                  hours: { type: integer }
                  from: { type: string, description: Inclusive UTC hour bucket (YYYY-MM-DDTHH). }
                  to: { type: string }
                  visits: { type: integer, description: Completed visits in the window. }
                  sections:
                    type: array
                    items:
                      type: object
                      required: [key, modules, views, medianDwellSeconds, medianDwellBucket, reach, depth]
                      properties:
                        key: { type: string, description: Nav node id. }
                        modules: { type: array, items: { type: string } }
                        views: { type: integer }
                        medianDwellSeconds: { type: [number, "null"] }
                        medianDwellBucket: { type: [string, "null"] }
                        reach: { type: [number, "null"], minimum: 0, maximum: 1 }
                        depth:
                          type: object
                          description: Raw scroll milestone counts.
                          properties:
                            "25": { type: integer }
                            "50": { type: integer }
                            "75": { type: integer }
                            "100": { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/cms/activity-names:
    get:
      tags: [CMS · Status]
      summary: Recorded activity names
      description: Distinct names from observed presence sessions, most-seen first, for the hidden-activities tag input.
      responses:
        "200":
          description: Names.
          content:
            application/json:
              schema:
                type: object
                required: [names]
                properties:
                  names:
                    type: array
                    items:
                      type: object
                      required: [name, category, sessions]
                      properties:
                        name: { type: string }
                        category: { $ref: "#/components/schemas/PresenceCategory" }
                        sessions: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ─────────────────────────────── CMS · Revisions ────────────────────────────
...
components:
  securitySchemes:
    cmsSession:
      type: apiKey
      in: cookie
      name: lg_session
      description: |
        Signed, http-only session cookie issued by the GitHub OAuth flow. This is the primary
        credential the CMS UI uses; requests are sent with credentials (cross-origin CORS is
        locked to the configured web origin).
    cmsBearer:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <CMS_TOKEN>` — a fallback for scripts and first-time setup. Prefer
        the OAuth session for interactive use.
...
  responses:
    Ok:
      description: Success.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Ok"
    BadRequest:
      description: The body, params, or query failed boundary validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: Missing or invalid CMS credential (session cookie or bearer token).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Ok:
      type: object
      description: A bare success acknowledgement.
      required: [ok]
      properties:
        ok:
          const: true
      additionalProperties: false
    Localized:
      type: object
      description: A human-authored string per locale. English is required; only `en` and `de` keys are accepted.
      required: [en]
      additionalProperties: false
      properties:
        en: { type: string }
        de: { type: string }
    PresenceCategory:
      type: string
      enum: [game, streaming, music, podcast, watching, custom]
...
    HeroStatusBody:
      type: object
      required: [verb, now]
      additionalProperties: false
      properties:
        verb: { $ref: "#/components/schemas/Localized" }
        now: { $ref: "#/components/schemas/Localized" }
...
    Error:
      type: object
      description: |
        The uniform error body. Every failure is mapped to this shape by one server-side error
        handler; internals and stack traces are never leaked.
      required: [error]
      properties:
        error:
          type: string
          description: A human-readable message safe to show a client.
      additionalProperties: false
...